+++ TransferEasy.Domain/TransactionEntry.cs
@@ -2,7 +2,7 @@
 
 public record TransactionEntry(int AccountId, Entry Entry);
 
-public record Entry(double Amount, EntryDirection Direction);
+public record Entry(decimal Amount, EntryDirection Direction);
 
 public enum EntryDirection
 {
